src/final_assignment/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
pydantic==2.5.2
psycopg2-binary==2.9.9
gunicorn==21.2.0
pytest==7.4.3
pytest-cov==4.1.0
requests==2.31.0